+++ apps/web/lib/inngest/index.ts
@@ -1,5 +1,3 @@
 export { inngest } from "./client";
-export { executeTask } from "./functions/execute-task";
 export { cleanupProjectSandbox } from "./functions/cleanup-project-sandbox";
 export { startSandbox, stopSandbox } from "./functions/session-sandbox";
-export { buildProject } from "./functions/build-project";